pyproject.toml
[tool.poetry]
name = "vtuber-scraper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "vtuber_scraper"}]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.5.0"
webdriver-manager = "^3.8.4"
google-api-python-client = "^2.65.0"
python-dotenv = "^0.21.0"
yt-dlp = "^2022.10.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
